Dynamically created sliders working
Found another way to handle the different sliders - a bit clumsy though.
Every control is pushed to the Controls array, and then assigned a serial number. This is then applied as a HTML attribute in the generated element, so that an event listener can find and invoke functions on the appropriate object.