OK further tweaks then
Still behaves WEIRDLY for low elementLength - think this is because the tiny masses at this scale are accelerated enormously by the pressure.
This is OK for now because the final element lengths might be of the order of 10cm at a minimum, rather than the 10mm garden we're playing in here.