Trying to make a better Pump
The pump should create a pressure difference across an interface by taking mass from one side and cramming it into the other.  However, the process should stall if the pressure difference becomes too great.  Some of these features are in place but the new Pump can happily outstrip its supply with little consequence...trying to find the fundamental reasons for this.