Updating Tank object
Tank works reasonably well - though suffers cavitation at the pump interface, so cannot supply continuous flow - UNREALISTIC as it should be able to meet any demand.

It would be nice to find a way to efficiently link the 3 valves that connect to the TankToPump, and 'hide' the controls for the valves indirectly controlled by the user.