Getting gravity to work
can now specify a z-position (posZ) for a given PipeElement. Differences in posZ across an Interface bring the weight of the water in the higher element into the force balance to a varying degree (less if the two PipeElements are at close to the same posZ).

Also threw in a 'displayScale' constant so we can now have big elementLengths without drawing off screen, which means METRES of hose without bogging down the sim with 1000s of elements.